requirements.txt
opencv-python~=4.10.0.84
numpy~=2.1.1
Flask~=3.0.3
requests~=2.32.3
pillow~=10.4.0